contrib/ecalproto/include/ecal/protobuf/ecal_proto_hlp.h synched with ecal/include/ecal/protobuf/ecal_proto_hlp.h contrib/ecalproto/include/ecal/protobuf/ecal_proto_dyn.h synched with ecal/include/ecal/protobuf/ecal_proto_dyn.h (header only now)
